Wire up ProfilePage and ProfileBloc. ProfileBloc instanciated in main.dart

+++ backend/controllers/v0/users/routes/auth.router.ts
@@ -87,7 +87,7 @@ router.post('/login', async (req: Request, res: Response) => {
     // Generate JWT
     const jwt = generateJWT(user);
 
-    res.status(200).send({ auth: true, token: jwt, user: user.short()});
+    res.status(200).send({ auth: true, token: jwt, user: user});
 });
 
 //register a new user
